+++ Source/Game/Game/Enemy.cpp
@@ -22,7 +22,7 @@ void Enemy::Update(float dt) {
         data.speedMin = 50;
         data.speedMax = 250;
         data.damping = 0.5f;
-        data.color = kiko::Color{ 1, 0.1, 0, 1 };
+        data.color = kiko::Color{ 1, (float)0.1, 0, 1 };
 
         kiko::Transform transform{ m_transform.position, 0, 1 };
         auto emitter = std::make_unique<kiko::Emitter>(transform, data);
